+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 <a name="0.9.2"></a>
 
-## [0.9.2](https://github.com/lars-sh/json-dom/compare/88a2a0fbea8b21b5ff0d3973a49d8494e089dd3a...master) (2024-03-20)
+## [0.9.2](https://github.com/lars-sh/json-dom/compare/88a2a0fbea8b21b5ff0d3973a49d8494e089dd3a...b967cbc1c1f87ce79033242ab027ead1fdb9b64c) (2024-03-20)
 
 Download at [Maven Search](https://search.maven.org/artifact/de.lars-sh/json-dom/0.9.2/jar)